extract get_hourly_usage-specific tests into their own test cases
